updated requir.. and made test folder
