+++ app.py
@@ -1,4 +1,4 @@
-from formapp import create_app
+from formapp.app import create_app
 
 app = create_app()
