+++ requirements.txt
@@ -1,20 +1,15 @@
-alembic==1.11.1
 blinker==1.6.2
 Bootstrap-Flask==2.3.0
 click==8.1.6
-colorama==0.4.6
 Flask==2.3.2
 Flask-Login==0.6.2
-Flask-MySQLdb==1.0.1
 Flask-SQLAlchemy==3.0.5
 Flask-WTF==1.1.1
 greenlet==2.0.2
 itsdangerous==2.1.2
 Jinja2==3.1.2
-Mako==1.2.4
 MarkupSafe==2.1.3
 mysql-connector-python==8.1.0
-mysqlclient==2.2.0
 protobuf==4.21.12
 python-dotenv==1.0.0
 SQLAlchemy==2.0.19
