+++ tests/test-requirements.txt
@@ -1,0 +1,3 @@
+Flask-Testing==0.8.1
+pytest==6.2.5
+pytest-cov==2.12.1
